see Changelog (#102)

+++ src/db/db_game.cpp
@@ -2824,7 +2824,7 @@ Game::setUndoLevel(unsigned level, unsigned combinePredecessingMoves)
 	if (m_undoList.size() > m_maxUndoLevel)
 	{
 		size_t n = m_undoList.size() - m_maxUndoLevel;
-		size_t r = mstl::min(m_undoList.size() - m_undoIndex, n);
+		size_t r = mstl::min(size_t(m_undoList.size() - m_undoIndex), n);
 		size_t k = m_undoList.size() - r;
 
 		// firstly erase redo's
